1.  Moved utils folder and related code inside proofofwork. 2.  Updated proofofwork.go imports.
