+++ blockchain/proofofwork/utils/utils.go
